requirements.txt
asgiref==3.5.0
blis==0.7.6
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cymem==2.0.6
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
gunicorn==20.1.0
idna==3.3
Jinja2==3.0.3
langcodes==3.3.0
MarkupSafe==2.1.0
murmurhash==1.0.6
numpy==1.22.3
packaging==21.3
pathy==0.6.1
preshed==3.0.6
psycopg2==2.9.3
pydantic==1.8.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
smart-open==5.2.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
tqdm==4.63.0
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
whitenoise==6.0.0
